Minor consistency fix in Text::GetFont(), the filename version. Move UIComponent getters to implementation to avoid requiring class definitions.

+++ Source/Urho3D/UI/Text.h
@@ -94,7 +94,7 @@ public:
     virtual void OnIndentSet();
 
     /// Set font by looking from resource cache by name and font size. Return true if successful.
-    bool SetFont(const String& fontName, int size = DEFAULT_FONT_SIZE);
+    bool SetFont(const String& fontName, float size = DEFAULT_FONT_SIZE);
     /// Set font and font size. Return true if successful.
     bool SetFont(Font* font, float size = DEFAULT_FONT_SIZE);
     /// Set font size only while retaining the existing font. Return true if successful.
